Refactor import paths in mock modules to align with new modular structure
This commit updates the import statements in various mock modules to reflect the new organization under the `realtime` directory. The changes ensure that all references to `MockRealtimeClient` and `MockResponseConfig` are consistent with the refactored structure, enhancing code maintainability and clarity. Additionally, backward compatibility is maintained through updated import paths in test files, ensuring existing tests continue to function without modification.
