+++ opusagent/mock/realtime/generators.py
